test: add shared runtime fixtures
